Code works

+++ eproduct_api_rest/src/main/java/com/edu/unicauca/eproducts/eproduct_api_rest/facade/services/EproductServiceImpl.java
@@ -1,5 +1,9 @@
 package com.edu.unicauca.eproducts.eproduct_api_rest.facade.services;
 
+import com.edu.unicauca.eproducts.eproduct_api_rest.controllers.exceptionsController.exceptionsEstructure.ErrorCode;
+import com.edu.unicauca.eproducts.eproduct_api_rest.controllers.exceptionsController.exceptionsLocal.EntidadNoExisteException;
+import com.edu.unicauca.eproducts.eproduct_api_rest.controllers.exceptionsController.exceptionsLocal.EntidadYaExisteException;
+import com.edu.unicauca.eproducts.eproduct_api_rest.controllers.exceptionsController.exceptionsLocal.ReglaNegocioException;
 import com.edu.unicauca.eproducts.eproduct_api_rest.dataAccess.models.EproductEntity;
 import com.edu.unicauca.eproducts.eproduct_api_rest.dataAccess.repositories.EproductRepository;
 import com.edu.unicauca.eproducts.eproduct_api_rest.facade.DTO.EproductDTO;
@@ -30,36 +34,73 @@ public class EproductServiceImpl implements IEproductService{
 
     @Override
     public EproductDTO findById(Integer id) {
-        EproductEntity eproductEntity = this.eproductRepository.findById(id);
-        return this.modelMapper.map(eproductEntity, EproductDTO.class);
+        if(this.eproductRepository.findById(id) == null){
+            throw new EntidadNoExisteException("El producto no existe, no se puede buscar");
+        } else {
+            return this.modelMapper.map(this.eproductRepository.findById(id), EproductDTO.class);
+        }
     }
 
     @Override
     public EproductDTO save(EproductDTO eproduct) {
-        float precioFinal = calcularPrecioFinal(eproduct);
-        eproduct.setPrecioFinal(precioFinal);
-        EproductEntity eproductEntity = this.modelMapper.map(eproduct, EproductEntity.class);
-        eproductEntity.setCreateAt(new Date());
-        EproductEntity eproductEntitySaved = this.eproductRepository.save(eproductEntity);
-        return this.modelMapper.map(eproductEntitySaved, EproductDTO.class);
+        if(this.eproductRepository.existByName(eproduct.getName())){
+            throw new EntidadYaExisteException("El producto ya existe, no se puede guardar");
+        } else if(eproduct.getPrecioBase() < 0){
+            throw new ReglaNegocioException("El precio base no puede ser negativo");
+        } else if(eproduct.getMarca().equals("")){
+            throw new ReglaNegocioException("Marca no valida");
+        } else if(eproduct.getCategoria().equals("")){
+            throw new ReglaNegocioException("Categoria no valida");
+        } else if (eproduct.getPrecioBase() < 50) {
+            throw new ReglaNegocioException("El precio base no puede ser menor a 50");            
+        } else{            
+            float precioFinal = calcularPrecioFinal(eproduct);
+            eproduct.setPrecioFinal(precioFinal);
+            EproductEntity eproductEntity = this.modelMapper.map(eproduct, EproductEntity.class);
+            eproductEntity.setCreateAt(new Date());
+            EproductEntity eproductEntitySaved = this.eproductRepository.save(eproductEntity);
+            return this.modelMapper.map(eproductEntitySaved, EproductDTO.class);
+        }
     }
 
     @Override
     public EproductDTO update(Integer id, EproductDTO eproduct) {
-        float precioFinal = calcularPrecioFinal(eproduct);
-        eproduct.setPrecioFinal(precioFinal);
-        EproductEntity eproductEntity = this.modelMapper.map(eproduct, EproductEntity.class);
-        EproductEntity eproductEntityUpdated = this.eproductRepository.update(id, eproductEntity);
-        if (eproductEntityUpdated == null) {
-            return null;
+        if (this.eproductRepository.findById(id) == null) {
+            throw new EntidadNoExisteException("El producto no existe, no se puede actualizar");            
+        } else if(eproduct.getPrecioBase() < 0){
+            throw new ReglaNegocioException("El precio base no puede ser negativo");
+        } else if(eproduct.getMarca().equals("")){
+            throw new ReglaNegocioException("Marca no valida");
+        } else if(eproduct.getCategoria().equals("")){
+            throw new ReglaNegocioException("Categoria no valida");
+        } else if (eproduct.getPrecioBase() < 50) {
+            throw new ReglaNegocioException("El precio base no puede ser menor a 50");            
+        } else{
+            String nameExist = this.eproductRepository.findById(id).getName();
+            String nameNew = eproduct.getName();
+            if(!nameExist.equals(nameNew) && this.eproductRepository.existByName(nameNew)){
+                throw new EntidadYaExisteException("El producto ya existe, no se puede actualizar");
+            }else{
+                float precioFinal = calcularPrecioFinal(eproduct);
+                eproduct.setPrecioFinal(precioFinal);
+                EproductEntity eproductEntity = this.modelMapper.map(eproduct, EproductEntity.class);
+                EproductEntity eproductEntityUpdated = this.eproductRepository.update(id, eproductEntity);
+                if (eproductEntityUpdated == null) {
+                    return null;
+                }
+                return this.modelMapper.map(eproductEntityUpdated, EproductDTO.class);
+            }
         }
-        return this.modelMapper.map(eproductEntityUpdated, EproductDTO.class);
     }
 
 
     @Override
     public boolean delete(Integer id) {
-        return this.eproductRepository.delete(id);
+        if (this.eproductRepository.findById(id) == null) {
+            throw new EntidadNoExisteException("El producto no existe, no se puede eliminar");            
+        } else {
+            return this.eproductRepository.delete(id);
+        }
     }
 
     public float calcularPrecioFinal(EproductDTO eproduct) {
